Completes ex. 6.6, adds action creator functions
This commit adds two action creator functions to
reducers/anecdoteReducer.js - yoVote, to create the action object for
voting on anecdotes, and addAnecdote, to create the action object for
adding a new anecdote. Both functions are imported into the App
component, which is modified to use them in the vote buttons and the
form submission handler function.