Completes ex. 6.7, adds component for new anecdote
This commit moves the functionality for creating a new anecdote from
App.js into a new component in components/AnecdoteForm.js. This new
component contains all of the logic for creating the new anecdote
that was previously a part of App.js; App.js now simply renders this
component without passing it any props.