Completes ex 6.8, adds anecdote list component
This commit moves the rendering of the anecdote list from App.js into
components/AnecdoteList.js. The new AnecdoteList component makes use
of react-redux useSelector to get the list of anecdotes, sorts them
by votes, and then renders each using a new Anecdote component
defined in components/Anecdote.js. The Anecdote component takes an
anecdote object as a prop and renders the anecdote, its number of
likes, and a like button using react-redux useDispatch.

I opted to separate out the Anecdote render into its own js file
instead of defining the component in the same file as AnecdoteList
as it felt like a cleaner split to separate into sub-components. This
requires passing data as props, which the redux lessons in FSO may be
trying to get away from (not entirely clear at the moment), but my
instincts are currently to separate into smaller distinct components.