Completes ex 6.9, adds filter for anecdotes
This commit adds a filter to the anecdote list that displays only
anecdotes that contain the entered filter string, case-insensitive.
This is implemented by adding a additional reducer in
reducers/filterReducer.js, and using CombineReducers in index.js to
enable complex state.

The filter is added in a new Filter component defined in
components/Filter.js. The filtering is done in the AnecdoteList
component, by use of the array filter function applied to the
anecdotes stored in state using the filter stored in state, via the
useSelector function.