Completes ex 6.12&13, adds notification component
This commit adds a notification component to the project. It appears
at the top of the anecdote list after a vote is recorded or a new
anecdote is added, with the information about what happened. Only the
most recent action is displayed, and the notification times out 5
seconds after the last action. This is enabled through Redux Toolkit
createSlice and configureStore, with functions notify to add a
notification and removeNotification to remove a notification.

The notification store is configured as an object with two members,
one being the count of notifications the other the content of the
most recent notification - either a string, or null if there is no
notification to be displayed. When notify is called the count is
incremented and the string is updated to the payload string. When
removeNotification is called, if the count is greater than 1 it is
decremented, and if less than or equal to 1 the count is set to 0 and
the content is set to null.

The Notification component returns an empty div if the content part
of the notification store is null.