Completes ex 6.14, gets initial state from server
This commit adds axios and json-server, and modifies the anecdotes
application to initially populate with the data saved in db.json.
Communication with the server is handled in services/anecdotes.js.

The initial population is handled by a new anecdote slicer action
overwriteAnecdoteState, which replaces the entire state with the
payload it is called with. I don't like having such an easily
accessed function that does this, it seems unsafe. However, the FSO
lesson leading into this exercise indicated to do it this way for
the entirely legitimate reason that it is much more efficient than
adding each individual anecdote in the database one at a time.
Hopefully later on in the FSO course there will be some further
refinement on this process.