Completes ex 6.16, inits store w/ action creator
This commit moves the initialization of the store from the backend DB
to an asyncronous action creator defined in reducers/anecdoteReducer.
This eliminates the export of overwriteAnecdoteState - instead, the
store is initialized through the action creator initializeAnecdotes,
which gets the anecdotes stored in the DB and calls
overwriteAnecdoteState to populate the store; initializeAnecdotes is
exported, and called in App.js to initialize the anecdotes app.