Completes ex 6.17, uses action creator for add
This commit moves creation of new anecdotes into an asyncronous
action creator in anecdoteReducer. The anecdoteSlice action
addAnecdote is renamed to appendAnecdote, and
reducers/anecdoteReducer now exports an async action creator as
addAnecdote. This action creator takes the content of the anecdote as
a parameter, POSTs it to the database via services/anecdotes.js, then
adds the returned object to the state via appendAnecdote.