Completes ex 6.18, uses action creator for votes
This commit modifies the functionality for voting and saving votes to
the backend so that it uses a Redux Thunk asyncronous action creator.
The anecdoteSlicer action yoVote is renamed to addVote, and
reducers/anecdoteReducer.js now exports an action creator named
yoVote which takes an anecdote id as its parameter. The action
creator uses getState and filter to find the anecdote with that id
(returning null if there are zero or more than one), calls the
addVote PUT function in services/anecdotes.js to update the DB, then
calls the addVote slicer action to update the store.