Passes entire anecdote when voting, vs just ID
This commit removes some unnecessary searching by id when voting for
an anecdote. The previous commit maintained the parameter format of
the yoVote function to just require id - this has some benefits when
working in a large code base (won't accidentally break things that
are calling the function), but doing so made the code less efficient
by requiring the action creator to find the entire anecdote object in
store from just the id, where the anecdote was fully available at the
one place yoVote is called. In this particular case, with this small
project, it is better to change the one call to pass the entire
object.