Completes ex 6.19, adds notification act'n creator
This commit creates a new action creator for notifications, which
includes the timeout functionality. The new action creator,
setNotification, takes two parameters - the contents of the
notification and the time for timeout in seconds.

Because of the possibility of overlapping timeouts (i.e., an earlier
notification is added with a longer timeout, such that it has not yet
timedout when a later notification does timeout), the reducer for
notifications had to be significantly reworked. The store for
notifications now consists of an array of objects which have the
notification's text ('content' member) and an id member. The notify
action now assigns the id for a new notification as 1 more than the
id of the last notification in the array store, and pushes the object
into the store - since push assigns the object to the end of the
array, this ensures the largest id of active notifications will
always be the last item in the array. If the array is empty, the id
is set to 0.

The removeNotification action now filters the notification array to
remove any notifications with an id matching the payload - based on
the notify action, ids should always be unique.

The setNotification action creator first calls notify, then uses
getState to find the id of the last notification in the store array,
and prepares a setTimeout to call removeNotification for that id
after the number of seconds indicated in the parameters.

Because of the possibility of overlapping timeouts, it was unclear
how best to handle displaying notifications - if only the latest
notification is displayed, and a later notification times out before
an earlier one, does that mean to revert the notification displayed
to the earlier one (which had since been overwritten)? Because this
is unclear, there is code for two possible Notification components in
components/Notification.js. Currently commented out is code for a
component that will display the most recent not-timed-out
notification, meaning overwritten notifications may later return (as
described above). Currently active is code that displays all
notifications sequentially in separate notification boxes, each of
which disappears when it times out.