Completes ex 6.21, implements adding anecdotes
This commit implements the functionality to add new anecdotes to the
server using React Query. It also renders the newly added anecdotes.

Note: the exercise states:

  Note that the content of the anecdote must be at least 5 characters
  long, otherwise the server will reject the POST request. You don't
  have to worry about error handling now.

However, it does not appear that the intial configuration is set up
to reject anecdotes of less than 5 characters. It does not read as if
I am supposed to implement this myself as part of the exercise.