Completes ex 6.23, adds useReducer notifications
This commit enables notifications in the anecdotes application, using
React's useReducer functionality. The Reducer is defined in
NotificationContext.jsx, along with functions for accessing Context
and dispatch function.

The Context is an object that is either empty or contains the text of
the most recent notification and a count of how many notificaitons
have been added but not removed. Adding a notification ('NOTIFY')
changes the text to the payload of the dispatch call and increments
the count, or sets to 1 if the Context was previously empty (the
logic for this checks if count >= 0 is true, in case something got
screwed up in the object and the object is not empty but the count is
not positive - this shouldn't happen, but better to be safe!).
Removing a notification ('REMOVE') decrements the count by 1 if the
count is greater than 1; if 1 or below, it sets the Context to an
empty object.

The function to access Context defined in NotificationContext.jsx
returns only the text portion of the Context. The count is not
designed to be accessable outside of the Reducer.

The dispatch is called by the handler functions for voting and adding
new anecdotes - it sets the payload as an appropriate notification,
then sets a 5 second timeout call to remove a notification. The
structure of the Reducer function ensures that the last notification
will be displayed for 5 seconds, and any earlier ones will be
overwritten by later ones if they are overwritten before timing out.