Completes ex. 6.24, adds error handling
This commit adds error handling for calls to the server via Query. It
adds a notification with the error message if either voting or adding
returns an error as an onError parameter of the mutation function.

This commit also moves the notifications for success into the
onSuccess parameter functions in the mutations, which is where they
probably should've been all along.